chore: 컨플릭트 해결을 위한 파일 확장자 수정
